Add files via upload
Note: In this plot, the green line is the decision boundary for the logistic model and the yellow line is the decision boundary for the least squares model; the logistic model is the predictor learned from logistic regression and the least squares model is the predictor learned from (ordinary least squares) linear regression. Recall that the decision boundary for a predictor is the hyperplane orthogonal to the predictor.