Create blankpage
hw3.py implements an SVM for linear classification. polynomialkernel_degreeoftwo, RBFkernel_sigma_equals_four, RBFkernel_sigma_equals_two, and RBFkernel_sigma_equals_one represent the contour plot of the SVM, trained on an XOR dataset, tested on the area [-5, 5] x [-5, 5] with the kernel being the polynomial kernel with degree 2, the RBF kernel with sigma = 4, the RBF kernel with sigma = 2, and the RBF kernel with sigma = 1, respectively.